+++ src/test/java/com/epam/reportportal/cucumber/integration/feature/EmptySteps.java
@@ -2,6 +2,7 @@ package com.epam.reportportal.cucumber.integration.feature;
 
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
+import cucumber.api.java.en.When;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -16,5 +17,10 @@ public class EmptySteps {
 	@Then("I have another empty step")
 	public void i_have_another_empty_step() {
 		LOGGER.info("Inside 'I have another empty step'");
+	}
+
+	@When("I have one more empty step")
+	public void i_have_one_more_empty_step() {
+		LOGGER.info("I have one more empty step'");
 	}
 }