+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
-[Unreleased]
+## [Unreleased]
+### Fixed
+- A bug whe ambiguous item cases a Null Pointer Exception
+- Incorrect item type settings
+### Added
+- Nested steps support
 
 ## [5.0.0-BETA-12]
+### Added
+- multi-thread execution support
+- Test Case ID support
+### Fixed
+- codeRef reporting was added for every item in an item tree
+
+## [3.0.0]
+### New Features 
+- Migrate to 3x client generation
+- Asynchronous reporting
+
+## [2.6.1]
+### Changed 
+- access modifier to 'protected' for rootSuiteId in ScenarioReporter to allow inherit this reporter without unnecessary code duplication.
+
+## [2.6.0]
+### Released: 20 November 2016
+### New Features
+- Initial release to Public Maven Repositories